+++ domains/emissions/Emissions.API/API/Models/EnergiDataService/MixRecords.cs
@@ -2,7 +2,7 @@
 {
     public class MixRecord
     {
-        public double ShareTotal { get; }
+        public decimal ShareTotal { get; }
 
         public DateTime HourUTC { get; }
 
@@ -12,7 +12,7 @@
 
         public string ProductionType { get; }
 
-        public MixRecord(double shareTotal, DateTime hourUTC, string version, string priceArea, string productionType)
+        public MixRecord(decimal shareTotal, DateTime hourUTC, string version, string priceArea, string productionType)
         {
             ShareTotal = shareTotal;
             HourUTC = hourUTC;
